+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG - CapeJS
 
+## 1.0.0-beta.5 (Not yet released)
+
+### `Cape.Router`
+
+* Users can set the *root container* of components by passing it to the first
+  argument of the router's constructor. The default root container is `window`.
+
 ## 1.0.0-beta.4 (April 22, 2015)
 
 ### `Cape.Router`